+++ plugin/src/main/kotlin/fr/pickaria/Main.kt
@@ -1,12 +1,12 @@
 package fr.pickaria
 
 import co.aikar.commands.PaperCommandManager
+import fr.pickaria.model.datasources.openDatabase
 import fr.pickaria.vue.PingCommand
 import fr.pickaria.vue.chat.ChatFormat
 import fr.pickaria.vue.chat.Motd
 import fr.pickaria.vue.chat.PlayerJoin
 import org.bukkit.plugin.java.JavaPlugin
-
 
 class Main : JavaPlugin() {
 	override fun onEnable() {
@@ -22,5 +22,8 @@ class Main : JavaPlugin() {
 			it.registerEvents(ChatFormat(), this)
 			it.registerEvents(Motd(), this)
 		}
+
+		// Database
+		openDatabase(dataFolder.absolutePath + "/database")
 	}
 }
