--- plugins/database/src/main/resources/plugin.yml
@@ -1,7 +1,0 @@
-name: PickariaDatabase
-version: 1.0
-api-version: 1.19
-main: fr.pickaria.database.Main
-author: Quozul
-depend:
-  - PickariaLib
